Add filter table name on database backup

+++ core/lib/Thelia/Install/Database.php
@@ -17,6 +17,7 @@ use Propel\Runtime\Connection\ConnectionInterface;
 use Propel\Runtime\Connection\ConnectionWrapper;
 use Propel\Runtime\Propel;
 use Propel\Runtime\ServiceContainer\ServiceContainerInterface;
+use Thelia\Log\Tlog;
 
 /**
  * Class Database
@@ -99,6 +100,7 @@ class Database
      * @param  string                          $sql  SQL query
      * @param  array                           $args SQL request parameters (PDO style)
      * @throws \RuntimeException|\PDOException if something goes wrong.
+     * @return \PDOStatement
      */
     public function execute($sql, $args = array())
     {
@@ -113,6 +115,8 @@ class Database
         if ($success === false || $stmt->errorCode() != 0) {
             throw new \RuntimeException("Failed to execute SQL '$sql', arguments:" . print_r($args, 1).", error:".print_r($stmt->errorInfo(), 1));
         }
+
+        return $stmt;
     }
 
     /**
@@ -165,14 +169,24 @@ class Database
         $data[] = "\n\n";
 
         foreach ($tables as $table) {
-            $result = $this->connection->prepare('SELECT * FROM ' . $table);
-            $result->execute();
+            if (!preg_match("/^[\w_\-]+$/", $table)) {
+                Tlog::getInstance()->alert(
+                    sprintf(
+                        "Attempt to backup the db with this invalid table name: '%s'",
+                        $table
+                    )
+                );
+
+                continue;
+            }
+
+            $result = $this->execute('SELECT * FROM `' . $table . '`');
+
             $fieldCount = $result->columnCount();
 
             $data[] = 'DROP TABLE `' . $table . '`;';
 
-            $resultStruct = $this->connection->prepare('SHOW CREATE TABLE `' . $table . '`');
-            $resultStruct->execute();
+            $resultStruct = $this->execute('SHOW CREATE TABLE `' . $table . '`');
 
             $rowStruct = $resultStruct->fetch(PDO::FETCH_NUM);
